moved webtest files location
